fix : Model Entity Class 제거 및 User Entity Class 수정
