--- src/main/java/dnd/studyplanner/domain/member/model/Member.java
@@ -1,46 +1,0 @@
-package dnd.studyplanner.domain.member.model;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-/**
- * User Entity로 대체될 임시 클래스입니다.
- */
-@Getter
-@NoArgsConstructor
-@Entity
-public class Member {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column(nullable = false)
-	private String nickname;
-
-	@Column(nullable = false)
-	private String email;
-
-	@Enumerated(EnumType.STRING)
-	@Column(nullable = false)
-	private Role role;
-
-	@Builder
-	public Member(String nickname, String email, Role role) {
-		this.nickname = nickname;
-		this.email = email;
-		this.role = role;
-	}
-
-	public String getRoleKey() {
-		return this.role.getKey();
-	}
-}
