+++ src/main/java/dnd/studyplanner/domain/user/model/User.java
@@ -3,20 +3,17 @@ package dnd.studyplanner.domain.user.model;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 import dnd.studyplanner.domain.base.BaseEntity;
+import dnd.studyplanner.domain.member.model.Role;
+import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @Entity
+@NoArgsConstructor
 @Table(name = "users")
 public class User extends BaseEntity {
 
@@ -25,12 +22,17 @@ public class User extends BaseEntity {
 	@Column(name = "user_id")
 	private Long id;
 
-	// @Column(nullable = false) 테스트를 위해 임시로 미적용
+	@Column(nullable = false) // 테스트를 위해 임시로 미적용
 	private String userEmail;
 
-	// @Column(nullable = false) 테스트를 위해 임시로 미적용
+	// @Column(nullable = false) // 테스트를 위해 임시로 미적용
 	private String accessToken;
 
+	@Enumerated(EnumType.STRING)
+	@Column(nullable = false)
+	private Role role;
+
+	private String userNickName;
 	private String userName;
 	private int userAge;
 	private String userGender;
@@ -43,4 +45,32 @@ public class User extends BaseEntity {
 
 	@OneToMany(mappedBy = "solveUser", cascade = CascadeType.ALL)
 	private List<UserSolveQuestionBook> userSolveQuestionBooks = new ArrayList<>();
+
+	@Builder
+	public User(String userEmail, String accessToken, Role role, String userNickName, String userName, int userAge
+			, String userGender, String userRegion, String userProfileImageUrl) {
+		this.userEmail = userEmail;
+		this.accessToken = accessToken;
+		this.role = role;
+		this.userNickName = userNickName;
+		this.userName = userName;
+		this.userAge = userAge;
+		this.userGender = userGender;
+		this.userRegion = userRegion;
+		this.userProfileImageUrl = userProfileImageUrl;
+	}
+
+	public String getRoleKey() {
+		return this.role.getKey();
+	}
+
+
+	public void update(String userName, int userAge, String userGender, String userRegion, String userProfileImageUrl) {
+		this.userName = userName;
+		this.userAge = userAge;
+		this.userGender = userGender;
+		this.userRegion = userRegion;
+		this.userProfileImageUrl = userProfileImageUrl;
+	}
+
 }
