+++ src/main/java/dnd/studyplanner/repository/UserRepository.java
@@ -4,5 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dnd.studyplanner.domain.user.model.User;
 
+import java.util.Optional;
+
 public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findById(Long userId);
+
+    Optional<User> findByUserEmail(String email);
 }
